✨ Move the classes from Task to Category
Now that we have a API module and the Category is ready to be developed,
it is fair to move all the code to the correct module.
