+++ features/category/src/main/java/com/escodro/category/di/CategoryModule.kt
@@ -1,8 +1,16 @@
 package com.escodro.category.di
 
+import com.escodro.category.mapper.CategoryMapperImpl
+import com.escodro.category.presentation.CategoryListViewModelImpl
+import com.escodro.categoryapi.mapper.CategoryMapper
+import com.escodro.categoryapi.presentation.CategoryListViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 /**
  * Category dependency injection module.
  */
-val categoryModule = module { }
+val categoryModule = module {
+    viewModel<CategoryListViewModel> { CategoryListViewModelImpl(get(), get()) }
+    factory<CategoryMapper> { CategoryMapperImpl() }
+}
