+++ features/category/src/test/java/com/escodro/category/fake/LoadAllCategoriesFake.kt
@@ -1,4 +1,4 @@
-package com.escodro.task.presentation.fake
+package com.escodro.category.fake
 
 import com.escodro.domain.model.Category
 import com.escodro.domain.usecase.category.LoadAllCategories
