+++ features/category/src/test/java/com/escodro/category/presentation/CategoryListViewModelTest.kt
@@ -1,10 +1,9 @@
-package com.escodro.task.presentation.detail
+package com.escodro.category.presentation
 
+import com.escodro.category.fake.FAKE_DOMAIN_CATEGORY_LIST
+import com.escodro.category.fake.LoadAllCategoriesFake
+import com.escodro.category.mapper.CategoryMapperImpl
 import com.escodro.categoryapi.presentation.CategoryState
-import com.escodro.task.mapper.CategoryMapperImpl
-import com.escodro.task.presentation.category.CategoryListViewModelImpl
-import com.escodro.task.presentation.fake.FAKE_DOMAIN_CATEGORY_LIST
-import com.escodro.task.presentation.fake.LoadAllCategoriesFake
 import com.escodro.test.CoroutineTestRule
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
